Upgrade to heroku-bouncer ~> 0.9.0
This version supports OmniAuth 2, which closes a CVE for CSRF. We use
the new `:login_path` option from heroku-bouncer to specify our own UI
to prompt the user to authenticate. This means adding views and view
rendering, but I've tried to keep all of it as simple/minimal as
possible. We also addressed some confusion around requests for the
favicon which were tripping me up in dev mode.